Fix tests for Go 1.17 and update CI accordingly
