+++ .github/workflows/test.yml
@@ -1,0 +1,24 @@
+name: test
+
+on:
+  push:
+    branches: [ 'main' ]
+  pull_request:
+    branches: [ 'main' ]
+
+jobs:
+  tests:
+    name: Run tests
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        go: [ '1.13.x', '1.14.x', '1.15.x', '1.16.x', '1.17.x' ]
+    steps:
+      - name: Set up Go ${{ matrix.go }}
+        uses: actions/setup-go@v2
+      - name: Check out code into the Go module directory
+        uses: actions/checkout@v2
+        with:
+          go-version: ${{ matrix.go }}
+      - name: Run Go tests
+        run: go test -v ./...
