+++ samlidp/util_go116_test.go
@@ -1,3 +1,6 @@
+//go:build !go1.17
+// +build !go1.17
+
 package samlidp
 
 import (
